Corrected deploy tasks of boot recipes.
do_deploy tasks are not image tasks, so they should *not*
deploy files directly to the ${DEPLOY_DIR_IMAGE} because
of yocto internal machinery as well as state cache problems.

What should be done is inherit do_deploy class and put files
to be deployed in ${DEPLOYDIR}. Not very intuitive but it is
like this. Please check [1] for more information.

http://www.yoctoproject.org/docs/2.1/mega-manual/mega-manual.html#shared-state